+++ packages/sdks-tests/src/specs/index.ts
@@ -12,6 +12,7 @@ import { CONTENT as image } from './image.js';
 import { CONTENT as dataBindings } from './data-bindings.js';
 import { CONTENT as dataBindingStyles } from './data-binding-styles.js';
 import { CONTENT as abTest } from './ab-test.js';
+import { CONTENT as symbolAbTest } from './symbol-ab-test.js';
 import {
   CONTENT as customBreakpoints,
   CONTENT_RESET as customBreakpointsReset,
@@ -48,6 +49,7 @@ const pages = {
   '/data-bindings': dataBindings,
   '/data-binding-styles': dataBindingStyles,
   '/ab-test': abTest,
+  '/symbol-ab-test': symbolAbTest,
   '/custom-breakpoints': customBreakpoints,
   '/reactive-state': reactiveState,
   '/element-events': elementEvents,