+++ packages/sdks/e2e/vue-3/pages/catch-all.page.js
@@ -1,0 +1,3 @@
+import App from '../src/App.vue';
+
+export default { Page: App };