+++ packages/sdks/e2e/vue-3/renderer/_default.page.server.js
@@ -1,0 +1,23 @@
+import 'cross-fetch/dist/node-polyfill.js';
+
+import { renderToString } from '@vue/server-renderer';
+import { escapeInject, dangerouslySkipEscape } from 'vite-plugin-ssr/server';
+import { createApp } from './app';
+
+export { render };
+export { passToClient };
+
+// See https://vite-plugin-ssr.com/data-fetching
+const passToClient = ['pageProps'];
+
+async function render(pageContext) {
+  const app = createApp(pageContext);
+  const appHtml = await renderToString(app);
+
+  return escapeInject`<!DOCTYPE html>
+    <html>
+      <body>
+        <div id="app">${dangerouslySkipEscape(appHtml)}</div>
+      </body>
+    </html>`;
+}